requirements.txt
psycopg2-binary==2.9.5
PySide6==6.4.0.1
PySide6-Addons==6.4.0.1
PySide6-Essentials==6.4.0.1
shiboken6==6.4.0.1
sql-formatter==0.6.2
